migrates project to use gradle and intellij plugins
