--- .idea/libraries/io_reactivex_rxjava_1_3_8.xml
@@ -1,14 +1,0 @@
-<component name="libraryTable">
-  <library name="io.reactivex:rxjava:1.3.8" type="repository">
-    <properties maven-id="io.reactivex:rxjava:1.3.8" />
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/io/reactivex/rxjava/1.3.8/rxjava-1.3.8.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/io/reactivex/rxjava/1.3.8/rxjava-1.3.8-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/io/reactivex/rxjava/1.3.8/rxjava-1.3.8-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
