--- .idea/misc.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_8" project-jdk-name="IntelliJ IDEA IU-181.4892.42" project-jdk-type="IDEA JDK">
-    <output url="file://$PROJECT_DIR$/out" />
-  </component>
-</project>
