--- src/FlutterI18nIcons.java
@@ -1,11 +1,0 @@
-import com.intellij.openapi.util.IconLoader;
-
-import javax.swing.*;
-
-public interface FlutterI18nIcons {
-
-    Icon EmptyFlag = IconLoader.getIcon("/icons/flags/flag_empty.png", FlutterI18nIcons.class);             // 16x16
-    Icon ArbFile = IconLoader.getIcon("/icons/arb_file.png", FlutterI18nIcons.class);                       // 16x16
-    Icon ArbRefreshAction = IconLoader.getIcon("/icons/arb_refresh.png", FlutterI18nIcons.class);           // 16x16
-
-}
