+++ src/main/kotlin/eu/long1/flutter/i18n/Register.kt
