+++ src/test/kotlin/LibraryTest.java
@@ -1,0 +1,12 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+import org.junit.Test;
+import static org.junit.Assert.*;
+
+public class LibraryTest {
+    @Test public void testSomeLibraryMethod() {
+        Library classUnderTest = new Library();
+        assertTrue("someLibraryMethod should return 'true'", classUnderTest.someLibraryMethod());
+    }
+}
